Initial nodelet conversion check-in
Initial check-in of changes needed to run
move_base as a nodelet rather than as a
node. Note that this check-in builds but
still does not run properly. Open issues
include:
1) Topic namespacing and remapping do not
yet work correctly, due to the way nodehandles
are used throughout the various adjacent
classes (e.g., costmap, planning).
2) There is a threading issue introduced.
A mutex is being unlocked inappropriately
by a different thread than the one that
created it.